@@ -33,12 +33,16 @@
     "temp": "^0.9.0"
   },
   "devDependencies": {
-    "@types/mocha": "^5.2.7",
-    "@types/node": "^12.6.8",
-    "chai": "4.2.0",
-    "glob": "7.1.4",
-    "mocha": "6.2.0",
-    "ts-node": "^8.3.0",
-    "typescript": "^3.5.3"
+    "@types/chai": "^4.2.12",
+    "@types/cross-spawn": "^6.0.0",
+    "@types/lodash": "^4.14.119",
+    "@types/mocha": "^5.2.5",
+    "@types/node": "^10.12.15",
+    "@types/temp": "^0.8.32",
+    "chai": "3.5.0",
+    "glob": "7.1.1",
+    "mocha": "5.1.1",
+    "ts-node": "^7.0.1",
+    "typescript": "^4.0.3"
   }
 }
