@@ -1,15 +1,14 @@
-var chai = require("chai")
-var path = require("path");
-var compiler = require(path.join(__dirname, "../src"));
-var childProcess = require("child_process");
-var _ = require("lodash");
-var temp = require("temp");
+import * as chai from "chai";
+import * as path from "path";
+import * as _ from "lodash";
 
-var expect = chai.expect;
+import * as compiler from "../src"
 
-var fixturesDir = path.join(__dirname, "fixtures");
+const expect = chai.expect;
 
-function prependFixturesDir(filename) {
+const fixturesDir = path.join(__dirname, "fixtures");
+
+function prependFixturesDir(filename: string) {
   return path.join(fixturesDir, filename);
 }
 
@@ -18,31 +17,17 @@ describe("#compile", function () {
   this.timeout(300000);
 
   it("reports errors on bad source", function (done) {
-    var opts = {
+    const opts = {
       verbose: true,
       cwd: fixturesDir
     };
-    var compileProcess = compiler.compile(prependFixturesDir("Bad.elm"), opts);
+    const compileProcess = compiler.compile(prependFixturesDir("Bad.elm"), opts);
 
-    compileProcess.on("exit", function (exitCode) {
-      var desc = "Expected elm make to have exit code 1";
+    compileProcess.on("exit", function (exitCode: number) {
+      const desc = "Expected elm make to have exit code 1";
       expect(exitCode, desc).to.equal(1);
       done();
     });
-  });
-
-  it("throws when given an unrecognized argument", function () {
-    var opts = {
-      foo: "bar",
-      output: "/dev/null",
-      verbose: true,
-      cwd: fixturesDir
-    };
-
-    expect(function () {
-      var compileProcess = compiler.compile(prependFixturesDir("Parent.elm"), opts);
-
-    }).to.throw();
   });
 });
 
@@ -50,41 +35,29 @@ describe("#compileToString", function () {
   // Use an epic timeout because Travis on Linux can be SUPER slow.
   this.timeout(600000);
 
-  it("adds runtime options as arguments", function () {
-    var opts = {
-      verbose: true,
-      cwd: fixturesDir,
-      runtimeOptions: ["-A128M", "-H128M", "-n8m"]
-    };
-
-    return expect(compiler
-      ._prepareProcessArgs("a.elm", opts)
-      .join(" ")).to.equal("make a.elm +RTS -A128M -H128M -n8m -RTS");
-  });
-
   it("reports errors on bad syntax", function () {
-    var opts = {
+    const opts = {
       verbose: true,
       cwd: fixturesDir
     };
-    var compilePromise = compiler.compileToString(prependFixturesDir("Bad.elm"), opts);
+    const compilePromise = compiler.compileToString(prependFixturesDir("Bad.elm"), opts);
 
-    return compilePromise.catch(function (err) {
+    return compilePromise.catch(function (err: Error) {
       expect(err).to.be.an('error');
       expect(String(err))
         .to.contain("Compilation failed")
-        .and.contain("PARSE ERROR");
+        .and.contain("MISSING EXPRESSION");
     });
   });
 
   it("reports type errors", function () {
-    var opts = {
+    const opts = {
       verbose: true,
       cwd: fixturesDir
     };
-    var compilePromise = compiler.compileToString(prependFixturesDir("TypeError.elm"), opts);
+    const compilePromise = compiler.compileToString(prependFixturesDir("TypeError.elm"), opts);
 
-    return compilePromise.catch(function (err) {
+    return compilePromise.catch(function (err: Error) {
       expect(err).to.be.an('error');
       expect(String(err))
         .to.contain("Compilation failed")
@@ -92,36 +65,17 @@ describe("#compileToString", function () {
     });
   });
 
-  it("Rejects the Promise when given an unrecognized argument like `yes`", function () {
-    var opts = {
-      foo: "bar",
-      verbose: true,
-      cwd: fixturesDir
-    };
-
-    var compilePromise = compiler.compileToString(prependFixturesDir("Parent.elm"), opts);
-
-    return new Promise(function (resolve, reject) {
-      return compilePromise.then(function () {
-        reject("Expected the compilation promise to be rejected due to the unrecognized compiler argument.");
-      }).catch(function () {
-        resolve();
-      });
-    });
-  });
-
-
   it("works when run multiple times", function () {
-    var opts = {
+    const opts = {
       verbose: true,
       cwd: fixturesDir
     };
 
-    var runCompile = function () {
-      var compilePromise = compiler.compileToString(prependFixturesDir("Parent.elm"), opts)
+    const runCompile = function () {
+      const compilePromise = compiler.compileToString(prependFixturesDir("Parent.elm"), opts)
 
-      return compilePromise.then(function (result) {
-        var desc = "Expected elm make to return the result of the compilation";
+      return compilePromise.then(function (result: string) {
+        const desc = "Expected elm make to return the result of the compilation";
         expect(result.toString(), desc).to.be.a('string');
       });
     };
@@ -130,24 +84,75 @@ describe("#compileToString", function () {
     // the compilations instead. For details, see https://github.com/elm/compiler/issues/1853.
     // This issue is tracked as https://github.com/rtfeldman/node-elm-compiler/issues/86.
     let promiseChain = Promise.resolve();
-    for (var i = 0; i < 10; i++) {
+    for (let i = 0; i < 10; i++) {
       promiseChain = promiseChain.then(() => runCompile());
     }
     return promiseChain;
   });
 
   it("handles output suffix correctly", function () {
-    var opts = {
+    const opts = {
       verbose: true,
       cwd: fixturesDir,
       output: prependFixturesDir("compiled.html"),
     };
 
     return compiler.compileToString(prependFixturesDir("Parent.elm"), opts)
-      .then(function (result) {
-        var desc = "Expected elm make to return the result of the compilation";
+      .then(function (result: string) {
+        const desc = "Expected elm make to return the result of the compilation";
         expect(result.toString(), desc).to.be.a('string');
       });
+  });
+});
+
+describe("#compileSync", function () {
+  // Use a timeout of 5 minutes because Travis on Linux can be SUPER slow.
+  this.timeout(300000);
+
+  it("succeeds on SimplestMain", function () {
+    const opts = {
+      verbose: true,
+      cwd: fixturesDir
+    };
+    const compileProcess = compiler.compileSync(prependFixturesDir("SimplestMain.elm"), opts) as any;
+
+    const exitCode = compileProcess.status;
+    const desc = "Expected elm make to have exit code 0";
+    expect(exitCode, desc).to.equal(0);
+  });
+
+  it("reports errors on bad source", function () {
+    const opts = {
+      verbose: true,
+      cwd: fixturesDir
+    };
+    const compileProcess = compiler.compileSync(prependFixturesDir("Bad.elm"), opts) as any;
+
+    const exitCode = compileProcess.status;
+    const desc = "Expected elm make to have exit code 1";
+    expect(exitCode, desc).to.equal(1);
+  });
+});
+
+describe("#compileToStringSync", function () {
+  it('returns string JS output of the given elm file', function () {
+    const opts = { verbose: true, cwd: fixturesDir };
+    const result = compiler.compileToStringSync(prependFixturesDir("Parent.elm"), opts);
+
+    expect(result).to.include("_Platform_export");
+  });
+
+  it('returns html output given "html" output option', function () {
+    const opts = {
+      verbose: true,
+      cwd: fixturesDir,
+      output: prependFixturesDir('compiled.html'),
+    };
+    const result = compiler.compileToStringSync(prependFixturesDir("Parent.elm"), opts);
+
+    expect(result).to.include('<!DOCTYPE HTML>');
+    expect(result).to.include('<title>Parent</title>');
+    expect(result).to.include("_Platform_export");
   });
 });
 
@@ -156,19 +161,36 @@ describe("#compileWorker", function () {
   this.timeout(300000);
 
   it("works with BasicWorker.elm", function () {
-    var opts = {
+    const opts = {
       verbose: true,
       cwd: fixturesDir
     };
-    var compilePromise = compiler.compileWorker(
+    const compilePromise = compiler.compileWorker(
       prependFixturesDir(""),
       prependFixturesDir("BasicWorker.elm"),
-      "BasicWorker"
     );
 
-    return compilePromise.then(function (app) {
-      app.ports.reportFromWorker.subscribe(function (str) {
+    return compilePromise.then(function (app: any) {
+      app.ports.reportFromWorker.subscribe(function (str: string) {
         expect(str).to.equal("it's alive!");
+      });
+    })
+  });
+
+  it("accepts arguments", function () {
+    const opts = {
+      verbose: true,
+      cwd: fixturesDir
+    };
+    const compilePromise = compiler.compileWorker(
+      prependFixturesDir(""),
+      prependFixturesDir("EchoWorker.elm"),
+      "echo"
+    );
+
+    return compilePromise.then(function (app: any) {
+      app.ports.reportFromWorker.subscribe(function (str: string) {
+        expect(str).to.equal("You said: echo");
       });
     })
   });
